only wait on the server in the scheduler, not the worker
